fix: resolve static test violations in compiled config tests
- Extract anonymous test doubles into named classes to fix PHPMD
  UnusedLocalVariable false positives in anonymous class methods
- Add phpcs:disable for useless override warning in Event\Config\Data
  (constructor provides default cacheId, matching existing pattern in
  Catalog\Attribute\Config\Data)
- All PHPCS errors/warnings and PHPMD violations resolved
